Move test_signals and views into their own sub mod
This refactors the tests slightly so that we can move
some of them into their specific search apps.
