+++ datahub/search/contact/test/test_views.py
@@ -1,0 +1,73 @@
+import pytest
+from rest_framework import status
+from rest_framework.reverse import reverse
+
+from datahub.company.test.factories import ContactFactory
+
+from datahub.core.test_utils import APITestMixin
+
+
+pytestmark = pytest.mark.django_db
+
+
+@pytest.fixture
+def setup_data():
+    """Sets up data for the tests."""
+    ContactFactory(first_name='abc', last_name='defg')
+    ContactFactory(first_name='first', last_name='last')
+
+
+class TestSearch(APITestMixin):
+    """Tests search views."""
+
+    def test_search_contact(self, setup_es, setup_data):
+        """Tests detailed contact search."""
+        setup_es.indices.refresh()
+
+        term = 'abc defg'
+
+        url = reverse('api-v3:search:contact')
+
+        response = self.api_client.post(url, {
+            'original_query': term,
+            'last_name': 'defg',
+        })
+
+        assert response.status_code == status.HTTP_200_OK
+        assert response.data['count'] == 1
+        assert len(response.data['results']) == 1
+        assert response.data['results'][0]['last_name'] == 'defg'
+
+    def test_search_contact_no_filters(self, setup_es, setup_data):
+        """Tests case where there is no filters provided."""
+        setup_es.indices.refresh()
+
+        url = reverse('api-v3:search:contact')
+        response = self.api_client.post(url, {})
+
+        assert response.status_code == status.HTTP_200_OK
+        assert len(response.data['results']) > 0
+
+    def test_search_contact_sort_by_last_name_desc(self, setup_es, setup_data):
+        """Tests sorting in descending order."""
+        ContactFactory(first_name='test_name', last_name='abcdef')
+        ContactFactory(first_name='test_name', last_name='bcdefg')
+        ContactFactory(first_name='test_name', last_name='cdefgh')
+        ContactFactory(first_name='test_name', last_name='defghi')
+
+        setup_es.indices.refresh()
+
+        term = 'test_name'
+
+        url = reverse('api-v3:search:contact')
+        response = self.api_client.post(url, {
+            'original_query': term,
+            'sortby': 'last_name:desc',
+        })
+
+        assert response.status_code == status.HTTP_200_OK
+        assert response.data['count'] == 4
+        assert ['defghi',
+                'cdefgh',
+                'bcdefg',
+                'abcdef'] == [contact['last_name'] for contact in response.data['results']]
